Applied classpath patches
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@1388 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/classpath/5.0/java/lang/Short.java
@@ -374,7 +374,7 @@ public final class Short extends Number implements Comparable<Short>
    * Reverse the bytes in val.
    * @since 1.5
    */
-  public static short reverseBytes(int val)
+  public static short reverseBytes(short val)
   {
     return (short) (((val >> 8) & 0xff) | ((val << 8) & 0xff00));
   }
